Refactor: Separate mixin
